requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Werkzeug==3.0.3
python-dotenv==1.0.1
pyserial==3.5
eventlet==0.36.1
gunicorn==22.0.0
